+++ ranger-r-package/ranger/tests/testthat/test_ranger.R
@@ -21,9 +21,9 @@ test_that("regression result is of class ranger with 14 elements", {
   expect_that(length(rg.reg), equals(14))
 })
 
-test_that("survival result is of class ranger with 15 elements", {
+test_that("survival result is of class ranger with 16 elements", {
   expect_that(rg.surv, is_a("ranger"))
-  expect_that(length(rg.surv), equals(15))
+  expect_that(length(rg.surv), equals(16))
 })
 
 #TODO: This would require the GenABEL package in suggests
@@ -133,6 +133,18 @@ test_that("Matrix interface works for classification", {
 test_that("Matrix interface prediction works for classification", {
   dat <- data.matrix(iris)
   rf <- ranger(dependent.variable.name = "Species", data = dat, write.forest = TRUE, classification = TRUE)
+  expect_that(predict(rf, dat), not(throws_error()))
+})
+
+test_that("Matrix interface works for Probability estimation", {
+  rf <- ranger(dependent.variable.name = "Species", data = data.matrix(iris), write.forest = TRUE, probability = TRUE)
+  expect_that(rf$treetype, equals("Probability estimation"))
+  expect_that(rf$forest$independent.variable.names, equals(colnames(iris)[1:4]))
+})
+
+test_that("Matrix interface prediction works for Probability estimation", {
+  dat <- data.matrix(iris)
+  rf <- ranger(dependent.variable.name = "Species", data = dat, write.forest = TRUE, probability = TRUE)
   expect_that(predict(rf, dat), not(throws_error()))
 })
 
@@ -574,6 +586,18 @@ test_that("If respect.unordered.factors=FALSE, regard characters as ordered", {
   expect_that(rf.char$prediction.error, equals(rf.fac$prediction.error))
 })
 
+test_that("maxstat splitting works for survival", {
+  rf <- ranger(Surv(time, status) ~ ., veteran, splitrule = "maxstat")
+  expect_that(rf, is_a("ranger"))
+})
+
+test_that("maxstat splitting, alpha out of range throws error", {
+  expect_that(ranger(Surv(time, status) ~ ., veteran, splitrule = "maxstat", alpha = -1), 
+              throws_error())
+  expect_that(ranger(Surv(time, status) ~ ., veteran, splitrule = "maxstat", alpha = 2), 
+              throws_error())
+})
+
 test_that("holdout mode holding out data with 0 weight", {
   weights <- rbinom(nrow(iris), 1, 0.5)
   rf <- ranger(Species ~ ., iris, num.trees = 5, importance = "permutation",  
@@ -604,3 +628,6 @@ test_that("holdout mode: no OOB prediction if no 0 weights", {
                holdout = TRUE, keep.inbag = TRUE)
   expect_that(all(is.na(rf$predictions)), is_true())
 })
+
+
+