+++ source/src/Forest/Forest.cpp
@@ -32,7 +32,7 @@
 #include <string>
 #include <ctime>
 #include <math.h>
-#ifndef WIN_R_BUILD
+#ifndef OLD_WIN_R_BUILD
 #include <thread>
 #include <chrono>
 #endif
@@ -47,8 +47,8 @@ Forest::Forest() :
     verbose_out(0), num_trees(DEFAULT_NUM_TREE), mtry(0), min_node_size(0), num_variables(0), num_independent_variables(
         0), seed(0), dependent_varID(0), num_samples(0), prediction_mode(false), memory_mode(MEM_DOUBLE), sample_with_replacement(
         true), memory_saving_splitting(false), splitrule(DEFAULT_SPLITRULE), predict_all(false), keep_inbag(false), sample_fraction(
-        1), holdout(false), num_threads(DEFAULT_NUM_THREADS), data(0), overall_prediction_error(0), importance_mode(
-        DEFAULT_IMPORTANCE_MODE), progress(0) {
+        1), holdout(false), alpha(DEFAULT_ALPHA), minprop(DEFAULT_MINPROP), num_threads(DEFAULT_NUM_THREADS), data(0), overall_prediction_error(
+        0), importance_mode(DEFAULT_IMPORTANCE_MODE), progress(0) {
 }
 
 Forest::~Forest() {
@@ -63,7 +63,7 @@ void Forest::initCpp(std::string dependent_variable_name, MemoryMode memory_mode
     std::string split_select_weights_file, std::vector<std::string>& always_split_variable_names,
     std::string status_variable_name, bool sample_with_replacement, std::vector<std::string>& unordered_variable_names,
     bool memory_saving_splitting, SplitRule splitrule, std::string case_weights_file, bool predict_all,
-    double sample_fraction, bool holdout) {
+    double sample_fraction, double alpha, double minprop, bool holdout) {
 
   this->verbose_out = verbose_out;
 
@@ -97,7 +97,7 @@ void Forest::initCpp(std::string dependent_variable_name, MemoryMode memory_mode
   // Call other init function
   init(dependent_variable_name, memory_mode, data, mtry, output_prefix, num_trees, seed, num_threads, importance_mode,
       min_node_size, status_variable_name, prediction_mode, sample_with_replacement, unordered_variable_names,
-      memory_saving_splitting, splitrule, predict_all, sample_fraction, holdout);
+      memory_saving_splitting, splitrule, predict_all, sample_fraction, alpha, minprop, holdout);
 
   if (prediction_mode) {
     loadFromFile(load_forest_filename);
@@ -131,7 +131,7 @@ void Forest::initCpp(std::string dependent_variable_name, MemoryMode memory_mode
   if (holdout && !case_weights.empty()) {
     size_t nonzero_weights = 0;
     for (auto& weight : case_weights) {
-      if (weight > 0 ){
+      if (weight > 0) {
         ++nonzero_weights;
       }
     }
@@ -152,14 +152,15 @@ void Forest::initR(std::string dependent_variable_name, Data* input_data, uint m
     std::vector<std::vector<double>>& split_select_weights, std::vector<std::string>& always_split_variable_names,
     std::string status_variable_name, bool prediction_mode, bool sample_with_replacement,
     std::vector<std::string>& unordered_variable_names, bool memory_saving_splitting, SplitRule splitrule,
-    std::vector<double>& case_weights, bool predict_all, bool keep_inbag, double sample_fraction, bool holdout) {
+    std::vector<double>& case_weights, bool predict_all, bool keep_inbag, double sample_fraction, double alpha,
+    double minprop, bool holdout) {
 
   this->verbose_out = verbose_out;
 
   // Call other init function
   init(dependent_variable_name, MEM_DOUBLE, input_data, mtry, "", num_trees, seed, num_threads, importance_mode,
       min_node_size, status_variable_name, prediction_mode, sample_with_replacement, unordered_variable_names,
-      memory_saving_splitting, splitrule, predict_all, sample_fraction, holdout);
+      memory_saving_splitting, splitrule, predict_all, sample_fraction, alpha, minprop, holdout);
 
   // Set variables to be always considered for splitting
   if (!always_split_variable_names.empty()) {
@@ -187,7 +188,7 @@ void Forest::init(std::string dependent_variable_name, MemoryMode memory_mode, D
     std::string output_prefix, uint num_trees, uint seed, uint num_threads, ImportanceMode importance_mode,
     uint min_node_size, std::string status_variable_name, bool prediction_mode, bool sample_with_replacement,
     std::vector<std::string>& unordered_variable_names, bool memory_saving_splitting, SplitRule splitrule,
-    bool predict_all, double sample_fraction, bool holdout) {
+    bool predict_all, double sample_fraction, double alpha, double minprop, bool holdout) {
 
   // Initialize data with memmode
   this->data = input_data;
@@ -202,7 +203,7 @@ void Forest::init(std::string dependent_variable_name, MemoryMode memory_mode, D
 
   // Set number of threads
   if (num_threads == DEFAULT_NUM_THREADS) {
-#ifdef WIN_R_BUILD
+#ifdef OLD_WIN_R_BUILD
     this->num_threads = 1;
 #else
     this->num_threads = std::thread::hardware_concurrency();
@@ -226,6 +227,8 @@ void Forest::init(std::string dependent_variable_name, MemoryMode memory_mode, D
   this->predict_all = predict_all;
   this->sample_fraction = sample_fraction;
   this->holdout = holdout;
+  this->alpha = alpha;
+  this->minprop = minprop;
 
   // Set number of samples and variables
   num_samples = data->getNumRows();
@@ -418,14 +421,15 @@ void Forest::grow() {
 
     trees[i]->init(data, mtry, dependent_varID, num_samples, tree_seed, &deterministic_varIDs, &split_select_varIDs,
         tree_split_select_weights, importance_mode, min_node_size, &no_split_variables, sample_with_replacement,
-        &is_ordered_variable, memory_saving_splitting, splitrule, &case_weights, keep_inbag, sample_fraction, holdout);
+        &is_ordered_variable, memory_saving_splitting, splitrule, &case_weights, keep_inbag, sample_fraction, alpha,
+        minprop, holdout);
   }
 
-  // Init variable importance
+// Init variable importance
   variable_importance.resize(num_independent_variables, 0);
 
-  // Grow trees in multiple threads
-#ifdef WIN_R_BUILD
+// Grow trees in multiple threads
+#ifdef OLD_WIN_R_BUILD
   progress = 0;
   clock_t start_time = clock();
   clock_t lap_time = clock();
@@ -444,7 +448,7 @@ void Forest::grow() {
   std::vector<std::thread> threads;
   threads.reserve(num_threads);
 
-  // Initailize importance per thread
+// Initailize importance per thread
   std::vector<std::vector<double>> variable_importance_threads(num_threads);
 
   for (uint i = 0; i < num_threads; ++i) {
@@ -477,7 +481,7 @@ void Forest::grow() {
 
 #endif
 
-  // Divide importance by number of trees
+// Divide importance by number of trees
   if (importance_mode == IMP_GINI) {
     for (auto& v : variable_importance) {
       v /= num_trees;
@@ -487,8 +491,8 @@ void Forest::grow() {
 
 void Forest::predict() {
 
-  // Predict trees in multiple threads and join the threads with the main thread
-#ifdef WIN_R_BUILD
+// Predict trees in multiple threads and join the threads with the main thread
+#ifdef OLD_WIN_R_BUILD
   progress = 0;
   clock_t start_time = clock();
   clock_t lap_time = clock();
@@ -521,14 +525,14 @@ void Forest::predict() {
 #endif
 #endif
 
-  // Call special functions for subclasses
+// Call special functions for subclasses
   predictInternal();
 }
 
 void Forest::computePredictionError() {
 
-  // Predict trees in multiple threads
-#ifdef WIN_R_BUILD
+// Predict trees in multiple threads
+#ifdef OLD_WIN_R_BUILD
   progress = 0;
   clock_t start_time = clock();
   clock_t lap_time = clock();
@@ -561,20 +565,20 @@ void Forest::computePredictionError() {
 
 void Forest::computePermutationImportance() {
 
-  // Compute tree permutation importance in multiple threads
-#ifdef WIN_R_BUILD
+// Compute tree permutation importance in multiple threads
+#ifdef OLD_WIN_R_BUILD
   progress = 0;
   clock_t start_time = clock();
   clock_t lap_time = clock();
 
-  // Initailize importance and variance
+// Initailize importance and variance
   variable_importance.resize(num_independent_variables, 0);
   std::vector<double> variance;
   if (importance_mode == IMP_PERM_BREIMAN || importance_mode == IMP_PERM_LIAW) {
     variance.resize(num_independent_variables, 0);
   }
 
-  // Compute importance
+// Compute importance
   for (size_t i = 0; i < num_trees; ++i) {
     trees[i]->computePermutationImportance(&variable_importance, &variance);
     progress++;
@@ -590,11 +594,11 @@ void Forest::computePermutationImportance() {
   std::vector<std::thread> threads;
   threads.reserve(num_threads);
 
-  // Initailize importance and variance
+// Initailize importance and variance
   std::vector<std::vector<double>> variable_importance_threads(num_threads);
   std::vector<std::vector<double>> variance_threads(num_threads);
 
-  // Compute importance
+// Compute importance
   for (uint i = 0; i < num_threads; ++i) {
     variable_importance_threads[i].resize(num_independent_variables, 0);
     if (importance_mode == IMP_PERM_BREIMAN || importance_mode == IMP_PERM_LIAW) {
@@ -615,7 +619,7 @@ void Forest::computePermutationImportance() {
   }
 #endif
 
-  // Sum thread importances
+// Sum thread importances
   variable_importance.resize(num_independent_variables, 0);
   for (size_t i = 0; i < num_independent_variables; ++i) {
     for (uint j = 0; j < num_threads; ++j) {
@@ -624,7 +628,7 @@ void Forest::computePermutationImportance() {
   }
   variable_importance_threads.clear();
 
-  // Sum thread variances
+// Sum thread variances
   std::vector<double> variance(num_independent_variables, 0);
   if (importance_mode == IMP_PERM_BREIMAN || importance_mode == IMP_PERM_LIAW) {
     for (size_t i = 0; i < num_independent_variables; ++i) {
@@ -649,7 +653,7 @@ void Forest::computePermutationImportance() {
   }
 }
 
-#ifndef WIN_R_BUILD
+#ifndef OLD_WIN_R_BUILD
 void Forest::growTreesInThread(uint thread_idx, std::vector<double>* variable_importance) {
   if (thread_ranges.size() > thread_idx + 1) {
     for (size_t i = thread_ranges[thread_idx]; i < thread_ranges[thread_idx + 1]; ++i) {
@@ -724,37 +728,37 @@ void Forest::computeTreePermutationImportanceInThread(uint thread_idx, std::vect
 void Forest::loadFromFile(std::string filename) {
   *verbose_out << "Loading forest from file " << filename << "." << std::endl;
 
-  // Open file for reading
+// Open file for reading
   std::ifstream infile;
   infile.open(filename, std::ios::binary);
   if (!infile.good()) {
     throw std::runtime_error("Could not read from input file: " + filename + ".");
   }
 
-  // Read dependent_varID and num_trees
+// Read dependent_varID and num_trees
   infile.read((char*) &dependent_varID, sizeof(dependent_varID));
   infile.read((char*) &num_trees, sizeof(num_trees));
 
-  // Read is_ordered_variable
+// Read is_ordered_variable
   readVector1D(is_ordered_variable, infile);
 
-  // Read tree data. This is different for tree types -> virtual function
+// Read tree data. This is different for tree types -> virtual function
   loadFromFileInternal(infile);
 
   infile.close();
 
-  // Create thread ranges
+// Create thread ranges
   equalSplit(thread_ranges, 0, num_trees - 1, num_threads);
 }
 
 void Forest::setSplitWeightVector(std::vector<std::vector<double>>& split_select_weights) {
 
-  // Size should be 1 x num_independent_variables or num_trees x num_independent_variables
+// Size should be 1 x num_independent_variables or num_trees x num_independent_variables
   if (split_select_weights.size() != 1 && split_select_weights.size() != num_trees) {
     throw std::runtime_error("Size of split select weights not equal to 1 or number of trees.");
   }
 
-  // Reserve space
+// Reserve space
   if (split_select_weights.size() == 1) {
     this->split_select_weights[0].resize(num_independent_variables);
   } else {
@@ -764,7 +768,7 @@ void Forest::setSplitWeightVector(std::vector<std::vector<double>>& split_select
   this->split_select_varIDs.resize(num_independent_variables);
   deterministic_varIDs.reserve(num_independent_variables);
 
-  // Split up in deterministic and weighted variables, ignore zero weights
+// Split up in deterministic and weighted variables, ignore zero weights
   for (size_t i = 0; i < split_select_weights.size(); ++i) {
 
     // Size should be 1 x num_independent_variables or num_trees x num_independent_variables
@@ -825,10 +829,10 @@ void Forest::setAlwaysSplitVariables(std::vector<std::string>& always_split_vari
   }
 }
 
-#ifdef WIN_R_BUILD
+#ifdef OLD_WIN_R_BUILD
 void Forest::showProgress(std::string operation, clock_t start_time, clock_t& lap_time) {
 
-  // Check for user interrupt
+// Check for user interrupt
   if (checkInterrupt()) {
     throw std::runtime_error("User interrupt.");
   }
@@ -853,7 +857,7 @@ void Forest::showProgress(std::string operation) {
   steady_clock::time_point last_time = steady_clock::now();
   std::unique_lock<std::mutex> lock(mutex);
 
-  // Wait for message from threads and show output if enough time elapsed
+// Wait for message from threads and show output if enough time elapsed
   while (progress < num_trees) {
     condition_variable.wait(lock);
     seconds elapsed_time = duration_cast<seconds>(steady_clock::now() - last_time);