+++ source/src/utility/ArgumentHandler.cpp
@@ -1,6 +1,6 @@
 /*-------------------------------------------------------------------------------
 This file is part of Ranger.
-    
+
 Ranger is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation, either version 3 of the License, or
@@ -14,13 +14,13 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with Ranger. If not, see <http://www.gnu.org/licenses/>.
 
-Written by: 
+Written by:
 
 Marvin N. Wright
 Institut für Medizinische Biometrie und Statistik
 Universität zu Lübeck
 Ratzeburger Allee 160
-23562 Lübeck 
+23562 Lübeck
 
 http://www.imbs-luebeck.de
 wright@imbs.uni-luebeck.de
@@ -36,9 +36,10 @@ wright@imbs.uni-luebeck.de
 
 ArgumentHandler::ArgumentHandler(int argc, char **argv) :
     caseweights(""), depvarname(""), fraction(1), holdout(false), memmode(MEM_DOUBLE), savemem(false), predict(""), splitweights(
-        ""), nthreads(DEFAULT_NUM_THREADS), predall(false), file(""), impmeasure(DEFAULT_IMPORTANCE_MODE), targetpartitionsize(
-        0), mtry(0), outprefix("ranger_out"), probability(false), splitrule(DEFAULT_SPLITRULE), statusvarname(""), ntree(
-        DEFAULT_NUM_TREE), replace(true), verbose(false), write(false), treetype(TREE_CLASSIFICATION), seed(0) {
+        ""), nthreads(DEFAULT_NUM_THREADS), predall(false), alpha(DEFAULT_ALPHA), minprop(DEFAULT_MINPROP), file(""), impmeasure(
+        DEFAULT_IMPORTANCE_MODE), targetpartitionsize(0), mtry(0), outprefix("ranger_out"), probability(false), splitrule(
+        DEFAULT_SPLITRULE), statusvarname(""), ntree(DEFAULT_NUM_TREE), replace(true), verbose(false), write(false), treetype(
+        TREE_CLASSIFICATION), seed(0) {
   this->argc = argc;
   this->argv = argv;
 }
@@ -49,7 +50,7 @@ ArgumentHandler::~ArgumentHandler() {
 int ArgumentHandler::processArguments() {
 
   // short options
-  char const *short_options = "A:C:D:F:HM:NP:S:U:Zac:f:hil::m:o:pr:s:t:uvwy:z:";
+  char const *short_options = "A:C:D:F:HM:NP:S:U:XZa:b:c:f:hil::m:o:pr:s:t:uvwy:z:";
 
   // long options: longname, no/optional/required argument?, flag(not used!), shortname
     const struct option long_options[] = {
@@ -64,9 +65,11 @@ int ArgumentHandler::processArguments() {
       { "predict",              required_argument,  0, 'P'},
       { "splitweights",         required_argument,  0, 'S'},
       { "nthreads",             required_argument,  0, 'U'},
+      { "predall",              no_argument,        0, 'X'},
       { "version",              no_argument,        0, 'Z'},
 
-      { "predall",              no_argument,  0, 'a'},
+      { "alpha",                required_argument,  0, 'a'},
+      { "minprop",              required_argument,  0, 'b'},
       { "catvars",              required_argument,  0, 'c'},
       { "file",                 required_argument,  0, 'f'},
       { "help",                 no_argument,        0, 'h'},
@@ -165,14 +168,42 @@ int ArgumentHandler::processArguments() {
       }
       break;
 
+    case 'X':
+      predall = true;
+      break;
+
     case 'Z':
       displayVersion();
       return -1;
       break;
 
-    // lower case options
+      // lower case options
     case 'a':
-      predall = true;
+      try {
+        double temp = std::stod(optarg);
+        if (temp < 0 || temp > 1) {
+          throw std::runtime_error("");
+        } else {
+          alpha = temp;
+        }
+      } catch (...) {
+        throw std::runtime_error(
+            "Illegal argument for option 'alpha'. Please give a value between 0 and 1. See '--help' for details.");
+      }
+      break;
+
+    case 'b':
+      try {
+        double temp = std::stod(optarg);
+        if (temp < 0 || temp > 0.5) {
+          throw std::runtime_error("");
+        } else {
+          minprop = temp;
+        }
+      } catch (...) {
+        throw std::runtime_error(
+            "Illegal argument for option 'minprop'. Please give a value between 0 and 0.5. See '--help' for details.");
+      }
       break;
 
     case 'c':
@@ -247,6 +278,9 @@ int ArgumentHandler::processArguments() {
           break;
         case 3:
           splitrule = AUC_IGNORE_TIES;
+          break;
+        case 4:
+          splitrule = MAXSTAT;
           break;
         default:
           throw std::runtime_error("");
@@ -396,7 +430,7 @@ void ArgumentHandler::checkArguments() {
   }
 
   // Check splitrule
-  if ((splitrule == AUC || splitrule == AUC_IGNORE_TIES) && treetype != TREE_SURVIVAL) {
+  if ((splitrule == AUC || splitrule == AUC_IGNORE_TIES || splitrule == MAXSTAT) && treetype != TREE_SURVIVAL) {
     throw std::runtime_error("Illegal splitrule selected. See '--help' for details.");
   }
 
@@ -405,6 +439,10 @@ void ArgumentHandler::checkArguments() {
     throw std::runtime_error("Case weights required to use holdout mode.");
   }
 
+  // Unordered survival splitting only available for logrank splitrule
+  if (treetype == TREE_SURVIVAL && !catvars.empty() && splitrule != LOGRANK) {
+    throw std::runtime_error("Unordered splitting in survival trees only available for LOGRANK splitrule.");
+  }
 }
 
 void ArgumentHandler::displayHelp() {
@@ -455,7 +493,11 @@ void ArgumentHandler::displayHelp() {
   std::cout << "    " << "--splitrule RULE              Splitting rule:" << std::endl;
   std::cout << "    " << "                              RULE = 1: Gini for Classification, variance for Regression, logrank for Survival." << std::endl;
   std::cout << "    " << "                              RULE = 2: AUC for Survival, not available for Classification and Regression." << std::endl;
+  std::cout << "    " << "                              RULE = 3: AUC (ignore ties) for Survival, not available for Classification and Regression." << std::endl;
+  std::cout << "    " << "                              RULE = 4: MAXSTAT for Survival, not available for Classification and Regression." << std::endl;
   std::cout << "    " << "                              (Default: 1)" << std::endl;
+  std::cout << "    " << "--alpha VAL                   Significance threshold to allow splitting (MAXSTAT splitrule only)." << std::endl;
+  std::cout << "    " << "--minprop VAL                 Lower quantile of covariate distribtuion to be considered for splitting (MAXSTAT splitrule only)." << std::endl;
   std::cout << "    " << "--caseweights FILE            Filename of case weights file." << std::endl;
   std::cout << "    " << "--holdout                     Hold-out mode. Hold-out all samples with case weight 0 and use these for variable " << std::endl;
   std::cout << "    " << "                              importance and prediction error." << std::endl;
@@ -482,13 +524,13 @@ void ArgumentHandler::displayVersion() {
   std::cout << "Ranger version: " << RANGER_VERSION << std::endl;
   std::cout << std::endl;
   std::cout << "Please cite Ranger: " << std::endl;
-  std::cout << "Marvin N. Wright and .. (2014). Ranger. Journal." << std::endl;
+  std::cout << "Wright, M. N. & Ziegler, A. (2016). ranger: A Fast Implementation of Random Forests for High Dimensional Data in C++ and R. Journal of Statistical Software, in press." << std::endl;
   std::cout << std::endl;
   std::cout << "BibTeX:" << std::endl;
   std::cout << "@Article{," << std::endl;
-  std::cout << "    title = {Ranger}" << std::endl;
-  std::cout << "    author = {Marvin N. Wright and ..}," << std::endl;
-  std::cout << "    journal = {Journal}," << std::endl;
-  std::cout << "    year = {2014}," << std::endl;
+  std::cout << "    title = {ranger: {{A}} fast implementation of random forests for high dimensional data in {{C}}++ and {{R}}}," << std::endl;
+  std::cout << "    author = {Wright, Marvin N. and Ziegler, Andreas}," << std::endl;
+  std::cout << "    journal = {Journal of Statistical Software}," << std::endl;
+  std::cout << "    year = {2016}," << std::endl;
   std::cout << "}" << std::endl;
 }